Fix boarding issue
Because train movements and passenger movement are done sequentially
passengers will already see the train location *at the end of the
round*.  Therefore, trains that want to stay at the station can't be
marked boardable immediately (because the won't be boardable until start
of the next round) but are marked as 'will become boardable at end of
round' (e.g. after passenger movements are done).